Allow grid concept to handle missing entity types
The grid allows to have some entities of the grid missing for specific codimensions. However, the current concept check requires that all entities are present. This this is fixed by checking that the grid has this capability, and for index sets and entities, if the Codim<cc> traits are available.