[YASP] partially revert 33b2a7d8
beside the storage of the globale domain size, 33b2a7d8 introduced a
breaking change in the YaspGrid geometry computations if working on
periodic grids.

There is no documentation about the precise semantics after the
change, it breaks code predating 2015 and it violates the documentate
behavior of YaspGrid parallel information. In the currentstate it is
really hard to have parallel periodic meshes, without special casing
the sequential case.

Actually this change should never have been merged.

Reverting this commit fixes #195