[test] generalize checkIdSet
The Id check assumed, that there is an injective mapping between *all*
entities and their Ids. In the periodic case this might not always be
true, but only hold for interior cells.

The point is there we might have overlap or ghost cells corresponding
to a periodically mapped cell on the same node.